requirements.txt
coloredlogs==15.0.1
PyYAML==6.0.1
webex-bot==0.4.0
prettytable==3.8.0